refactor(pkg): Rename `rm` to `remove` to be more consistent
